Update termlist page build script and build 2020-08-23 version
The termlist page build script was modified to include the modification date and examples, which are now included in the new versions of borrowed Darwin Core terms. The recently released updated versions of Darwin Core terms have replaced the obsolete versions. So the DwC term definitions and recommendations are now in harmony with Darwin Core itself.