Add permalinks for all versioned pages
Currently many of the pages had the same permalink as the main page and were therefore inaccessible. Permalink is /pagename/yyyy-mm-dd. This is similar to Darwin Core, where pages are served at:

/pagename/yyyy-mm-dd
/pagename/yyyy-mm-dd.html
but not
/pagename/yyyy-mm-dd/ (closing slash)

cc @baskaufs these permalinks could be removed if pages are structured in directories, as you suggest in #250